use preload trick to work with new electron version
